Update LICENSE file
The current LICENSE file in this project has been altered to include JetBrains as the copyright owner. However, the Apache 2.0 license file should not be altered in any way and should be included in repos verbatim. The Appendix at the end is a template for people to apply the Apache License to their work (in the form of license headers in source code files), instead of a declaration of the owner of the current repo.